Combined Encoder Period and Count PIO Programs
The servo is using one of the PIOs, so there isn't
enough room to have 4 separate state machines
running for both the Encoder tick counter and the
encoder period counter.

This change replaces the encoder PIO program with
one that will output the time between encoder pulses
to its RX FIFO.  Instead of having the PIO program
keep track of the encoder tick count, the robot
code will count the number of items it pulls from
the RX FIFO (there will be one for each encoder
tick) and keep track of the count itself.

Only the encoder period (time between pulses)
will be calculated in the PIO.